refactor: move Chip component to components package
