+++ ui/src/main/java/com/example/ui/screens/search/sections/filterDialog/FilterDialog.kt
@@ -29,12 +29,12 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import com.example.designsystem.R
-import com.example.designsystem.chip.Chip
 import com.example.designsystem.components.Icon
 import com.example.designsystem.components.IconButton
 import com.example.designsystem.components.Text
 import com.example.designsystem.components.buttons.ConfirmButton
 import com.example.designsystem.components.buttons.OutlinedButton
+import com.example.designsystem.components.chip.Chip
 import com.example.designsystem.theme.AflamiTheme
 import com.example.designsystem.theme.AppTheme
 import com.example.designsystem.utils.ThemeAndLocalePreviews
